👷(project) add dynamic CI configuration
To avoid having to always test unchanged application trays, we've
designed a generic CI workflow that auto-updates given changed/added
apps and only test those target applications when required.

Note that adding a new tray to this repository still requires to create
a target application vault for the `eugene` customer in the `ci`
environment.