✨(grafana) add post deployment hooks support
Some dashboards may require to be modified once deployed by injecting
grafana database values in their source.

Those hooks are scripts distributed along with compiled dashboards and
will be executed in a side-car container so that it can access to the
temporary volume used to provision dashboards. This requires
post-deployment hooks scripts to be idempotent.